Some ground work for a system editor (the main issue will be letting people choose images.. I've no idea how to make that useable considering the sheer amount of possible images)
I'll need to go and implement the login soon, that's not much work. With basic networking functional and successfully broadcasting systems, it should be a day of full work to fix it. Once I feel better, i'll give it a shot. I mean - an update loop to check every 15 seconds for changes, a basic player class, a basic ship class... the biggest challenge left for a playable prototype is the chat and the linking with the outer website. It's very tempting to simply let players click a ship and select their actions, but ISW traditionally had a sidebar for these things. Implementing this sidebar isn't impossible, it's just a bit more JSNI than I have used in the past, and it may cause unexpected problems. That's actually the same thing that may make the chat take some time.

NPC's will be simple - ill just move them randomly and clamp them in XXX distance, targetting anyone who comes too close (ignoring the clamp once done).

 I'll probably need to import some GDX portions just to make sure the math is consistent, better to do this before its needed then have to recode all the algorithms later.